Make RemotingTerminator non-FSM to avoid racy FSM log init. (#7967)
RemotingTerminator is initialized during Provider Start. FSM constructor
initializes the _log with `context.GetLogger()` which eventually leads into
`LogSource.FromActorRef()`. This method attempts to access
`Provider.DefaultAddress`, which is being initialized.

If Provider `Init()` loses the race, either its `RemoteInternals` is not set,
causing NRE, or if `Remoting.Start()` is not completed, the `_defaultAddress`
is not set and returns null.

This commit removes the FSM from `RemotingTerminator` and manually implements
it using `Become()` API.

Co-authored-by: Aaron Stannard <aaron@petabridge.com>